[ningpop] complete community backend

+++ community/views.py
@@ -1,6 +1,7 @@
 from django.shortcuts import render, get_object_or_404, redirect
 from .models import Community
 from .forms import CommunityForm
+from django.core.exceptions import PermissionDenied
 
 # Create your views here.
 
@@ -12,10 +13,12 @@ def community_list(request):
 # Create
 def community(request):
     if request.method == 'POST':
-        community_form = CommunityForm(request.POST)
+        username = Community.objects.create(user=request.user)
+        community_form = CommunityForm(request.POST, request.FILES, instance=username)
+        
         if community_form.is_valid():
             community_form.save()
-            return redirect('/community/')
+            return redirect('/communitylist/')
     else:
         community_form = CommunityForm()
     return render(request, 'community.html', {'community_form':community_form})
@@ -30,17 +33,21 @@ def community_detail(request, community_id):
 # Update
 def community_update(request, community_id):
     community_update = get_object_or_404(Community, pk=community_id)
-    if request.method == 'POST':
-        community_form = CommunityForm(request.POST, instance=community_update)
-        if community_form.is_valid():
-            community_form.save()
-            return redirect('/community/detail/'+str(community_id))
-    else:
-        community_form = CommunityForm(instance=community_update)
-    return render(request, 'community_update.html', {'community_form':community_form})
+    if request.user == community_update.user:
+        if request.method == 'POST':
+            community_form = CommunityForm(request.POST, instance=community_update)
+            if community_form.is_valid():
+                community_form.save()
+                return redirect('/community/'+str(community_id))
+        else:
+            community_form = CommunityForm(instance=community_update)
+        return render(request, 'community_update.html', {'community_form':community_form})
+    raise PermissionDenied # 권한없음 오류
 
 # Delete
 def community_delete(request, community_id):
     community = Community.objects.get(pk=community_id)
-    community.delete() 
-    return redirect('/community/')
+    if request.user == community.user:
+        community.delete()
+        return redirect('/communitylist/')
+    raise PermissionDenied # 권한없음 오류
